+++ src/Presentation/SmartStore.Web/Administration/Controllers/QuantityUnitController.cs
@@ -3,7 +3,6 @@ using System.Linq;
 using System.Web.Mvc;
 using SmartStore.Admin.Models.Directory;
 using SmartStore.Core.Domain.Directory;
-using SmartStore.Services;
 using SmartStore.Services.Directory;
 using SmartStore.Services.Localization;
 using SmartStore.Services.Security;
@@ -17,45 +16,20 @@ namespace SmartStore.Admin.Controllers
     [AdminAuthorize]
     public class QuantityUnitController :  AdminControllerBase
     {
-        #region Fields
-
         private readonly IQuantityUnitService _quantityUnitService;
         private readonly ILocalizedEntityService _localizedEntityService;
         private readonly ILanguageService _languageService;
-		private readonly ICommonServices _services;
-
-        #endregion
-
-        #region Constructors
 
         public QuantityUnitController(
 			IQuantityUnitService quantityUnitService,
             ILocalizedEntityService localizedEntityService, 
-            ILanguageService languageService,
-			ICommonServices services)
+            ILanguageService languageService)
         {
             _quantityUnitService = quantityUnitService;
             _localizedEntityService = localizedEntityService;
             _languageService = languageService;
-			_services = services;
         }
         
-        #endregion
-
-        #region Utilities
-
-        [NonAction]
-        public void UpdateLocales(QuantityUnit quantityUnit, QuantityUnitModel model)
-        {
-            foreach (var localized in model.Locales)
-            {
-                _localizedEntityService.SaveLocalizedValue(quantityUnit, x => x.Name, localized.Name, localized.LanguageId);
-                _localizedEntityService.SaveLocalizedValue(quantityUnit, x => x.Description, localized.Description, localized.LanguageId);
-            }
-        }
-
-        #endregion
-
         #region Methods
 
         public ActionResult Index()
@@ -65,20 +39,25 @@ namespace SmartStore.Admin.Controllers
 
         public ActionResult List()
         {
-			if (!_services.Permissions.Authorize(StandardPermissionProvider.ManageMeasures))
+            if (!Services.Permissions.Authorize(StandardPermissionProvider.ManageMeasures))
+            {
                 return AccessDeniedView();
+            }
 
-            var quantityUnitModel = _quantityUnitService.GetAllQuantityUnits().Select(x => x.ToModel()).ToList();
+            var quantityUnitModel = _quantityUnitService.GetAllQuantityUnits()
+                .Select(x => x.ToModel())
+                .ToList();
 
             var gridModel = new GridModel<QuantityUnitModel>
             {
                 Data = quantityUnitModel,
                 Total = quantityUnitModel.Count()
             };
+
             return View(gridModel);
         }
 
-        //ajax
+        // Ajax.
         public ActionResult AllQuantityUnits(string label, int selectedId)
         {
             var quantityUnits = _quantityUnitService.GetAllQuantityUnits();
@@ -104,11 +83,12 @@ namespace SmartStore.Admin.Controllers
 
         public ActionResult Create()
         {
-			if (!_services.Permissions.Authorize(StandardPermissionProvider.ManageMeasures))
+            if (!Services.Permissions.Authorize(StandardPermissionProvider.ManageMeasures))
+            {
                 return AccessDeniedView();
+            }
 
             var model = new QuantityUnitModel();
-            //locales
             AddLocales(_languageService, model.Locales);
             
             return View(model);
@@ -117,40 +97,46 @@ namespace SmartStore.Admin.Controllers
         [HttpPost, ParameterBasedOnFormName("save-continue", "continueEditing")]
         public ActionResult Create(QuantityUnitModel model, bool continueEditing)
         {
-			if (!_services.Permissions.Authorize(StandardPermissionProvider.ManageMeasures))
+            if (!Services.Permissions.Authorize(StandardPermissionProvider.ManageMeasures))
+            {
                 return AccessDeniedView();
+            }
 
             if (ModelState.IsValid)
             {
                 var quantityUnit = model.ToEntity();
 
                 _quantityUnitService.InsertQuantityUnit(quantityUnit);
-                //locales
+
                 UpdateLocales(quantityUnit, model);
 
                 NotifySuccess(T("Admin.Configuration.QuantityUnit.Added"));
                 return continueEditing ? RedirectToAction("Edit", new { id = quantityUnit.Id }) : RedirectToAction("List");
             }
 
-            //If we got this far, something failed, redisplay form
             return View(model);
         }
 
         public ActionResult Edit(int id)
         {
-            if (!_services.Permissions.Authorize(StandardPermissionProvider.ManageMeasures))
+            if (!Services.Permissions.Authorize(StandardPermissionProvider.ManageMeasures))
+            {
                 return AccessDeniedView();
+            }
 
             var quantityUnit = _quantityUnitService.GetQuantityUnitById(id);
             if (quantityUnit == null)
+            {
                 return RedirectToAction("List");
+            }
 
             var model = quantityUnit.ToModel();
 
             AddLocales(_languageService, model.Locales, (locale, languageId) =>
             {
                 locale.Name = quantityUnit.GetLocalized(x => x.Name, languageId, false, false);
-				locale.Description = quantityUnit.GetLocalized(x => x.Description, languageId, false, false);
+                locale.NamePlural = quantityUnit.GetLocalized(x => x.NamePlural, languageId, false, false);
+                locale.Description = quantityUnit.GetLocalized(x => x.Description, languageId, false, false);
             });
 
             return View(model);
@@ -159,12 +145,16 @@ namespace SmartStore.Admin.Controllers
         [HttpPost, ParameterBasedOnFormName("save-continue", "continueEditing")]
         public ActionResult Edit(QuantityUnitModel model, bool continueEditing)
         {
-			if (!_services.Permissions.Authorize(StandardPermissionProvider.ManageMeasures))
+            if (!Services.Permissions.Authorize(StandardPermissionProvider.ManageMeasures))
+            {
                 return AccessDeniedView();
+            }
 
             var quantityUnit = _quantityUnitService.GetQuantityUnitById(model.Id);
             if (quantityUnit == null)
+            {
                 return RedirectToAction("List");
+            }
 
             if (ModelState.IsValid)
             {
@@ -181,39 +171,57 @@ namespace SmartStore.Admin.Controllers
             return View(model);
         }
 
+        [HttpPost]
+        public ActionResult Save(FormCollection formValues)
+        {
+            if (!Services.Permissions.Authorize(StandardPermissionProvider.ManageMeasures))
+            {
+                return AccessDeniedView();
+            }
+
+            return RedirectToAction("List", "QuantityUnit");
+        }
+
         [HttpPost, ActionName("Delete")]
         public ActionResult DeleteConfirmed(int id)
         {
-			if (!_services.Permissions.Authorize(StandardPermissionProvider.ManageMeasures))
+            if (!Services.Permissions.Authorize(StandardPermissionProvider.ManageMeasures))
+            {
                 return AccessDeniedView();
+            }
 
             var quantityUnit = _quantityUnitService.GetQuantityUnitById(id);
             if (quantityUnit == null)
-                //No delivery time found with the specified id
+            {
                 return RedirectToAction("List");
+            }
 
             try
             {
-
                 _quantityUnitService.DeleteQuantityUnit(quantityUnit);
 
-                NotifySuccess(T("Admin.Configuration.QuantityUnits.Deleted"));
+                NotifySuccess(T("Admin.Common.TaskSuccessfullyProcessed"));
                 return RedirectToAction("List");
             }
-            catch (Exception exc)
+            catch (Exception ex)
             {
-                NotifyError(exc);
+                NotifyError(ex);
                 return RedirectToAction("Edit", new { id = quantityUnit.Id });
             }
         }
 
-        [HttpPost]
-        public ActionResult Save(FormCollection formValues)
-        {
-			if (!_services.Permissions.Authorize(StandardPermissionProvider.ManageMeasures))
-                return AccessDeniedView();
+        #endregion
 
-            return RedirectToAction("List", "QuantityUnit");
+        #region Utilities
+
+        private void UpdateLocales(QuantityUnit quantityUnit, QuantityUnitModel model)
+        {
+            foreach (var localized in model.Locales)
+            {
+                _localizedEntityService.SaveLocalizedValue(quantityUnit, x => x.Name, localized.Name, localized.LanguageId);
+                _localizedEntityService.SaveLocalizedValue(quantityUnit, x => x.NamePlural, localized.NamePlural, localized.LanguageId);
+                _localizedEntityService.SaveLocalizedValue(quantityUnit, x => x.Description, localized.Description, localized.LanguageId);
+            }
         }
 
         #endregion